+++ test/acceptance/smoke.acceptance.test.ts
@@ -1,11 +1,11 @@
 // tslint:disable no-console
 import ansis from 'ansis'
 import {expect} from 'chai'
+import {execa} from 'execa'
 import fs from 'fs-extra'
 import path from 'node:path'
 import {fileURLToPath} from 'node:url'
 
-import {x} from '../../scripts/utils/exec.js'
 import normalizeTableOutput from '../helpers/utils/normalize-table-output.js'
 import commandsOutput from './commands-output.js'
 
@@ -16,7 +16,8 @@ const bin = path.join(__dirname, '../../bin/run')
 
 function run(args = '') {
   console.log(`$ heroku ${args}`)
-  return x([bin, args].join(' '), {stdio: undefined})
+  // Use execa directly to capture output for test assertions
+  return execa([bin, args].join(' '), {shell: true})
 }
 
 // Smoke tests expect the CI account: heroku-cli@salesforce.com, app heroku-cli-ci-smoke-test-app,
